tests: file-separated tests
